Revert part of morgs' fix for current activity propagation.
This fixes "current-activity stealing" on Salut, which is bug#3328.

It turns out to be harmful to remove other buddies' current activity if it
matches an activity we just saw them leave, because during the join process,
we don't necessarily see everyone in the room straight away - so we may believe
that they momentarily left the room, then came back.

This should ideally be fixed in the connection managers, but a fix there is
much more involved.

Anyway, the other part of morgs' fix for current activity propagation
(removing *our* current activity if *we* leave) is necessary and sufficient,
and this one is redundant. The test cases in #3291 still work, except where
they're broken by #3294.